Optionally accumulate only non-reentrant zone time
Introduce a compile-time option which, when enabled (as it
is by default in this commit), causes the total zone time
(as displayed when "Self time" is unchecked) to reflect only
non-reentrant zone entries.  Thus a (direct or indirect)
recursive call will not contribute to the total zone time
(time is already being accounted to that zone on behalf of
the call which is already on the stack).

The _count_ of the number of times that a zone is entered
is not changed by this commit -- recursive calls are still
counted, regardless of the state of the new compile-time
option.